create user in InitializationFilter, move `getOrCreateUser` to parent Manager
