+++ infrastructures/openshift/pom.xml
@@ -128,6 +128,10 @@
         </dependency>
         <dependency>
             <groupId>org.eclipse.che.multiuser</groupId>
+            <artifactId>che-multiuser-api-authorization</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.eclipse.che.multiuser</groupId>
             <artifactId>che-multiuser-keycloak-server</artifactId>
         </dependency>
         <dependency>
